+++ IntegrationEngine.Tests/Api/Controllers/JobTypeControllerTest.cs
@@ -6,7 +6,6 @@ using Moq;
 using NUnit.Framework;
 using System;
 using System.Linq;
-using System.Collections;
 using System.Collections.Generic;
 
 namespace IntegrationEngine.Tests.Api.Controllers