+++ IntegrationEngine.Tests/EngineHostCompositionRootTest.cs
@@ -4,11 +4,16 @@ using Common.Logging.NLog;
 using IntegrationEngine.Api;
 using IntegrationEngine.Core.Configuration;
 using IntegrationEngine.Core.R;
+using IntegrationServer.IntegrationJobs.CarReport;
+using IntegrationServer.IntegrationJobs.SampleSqlReport;
 using IntegrationEngine.JobProcessor;
 using Microsoft.Practices.Unity;
 using Moq;
 using Nest;
 using NUnit.Framework;
+using IntegrationServer;
+using System.Reflection;
+using System.Collections.Generic;
 
 namespace IntegrationEngine.Tests
 {
@@ -51,9 +56,21 @@ namespace IntegrationEngine.Tests
             Subject.Container.RegisterType<IRabbitMQConfiguration, RabbitMQConfiguration>(configName,
                 new InjectionConstructor(new ResolvedParameter<IEngineConfiguration>(), configName));
 
-            Subject.SetupThreadedListenerManager();
+            Subject.SetupMessageQueueListenerManager();
 
-            Subject.Container.Resolve<IThreadedListenerManager>();
+            Assert.That(Subject.MessageQueueListenerManager, Is.Not.Null);
+        }
+
+        [Test]
+        public void ShouldStartMessageQueueListener()
+        {
+            var mockMessageQueueListenerManager = new Mock<IMessageQueueListenerManager>();
+            mockMessageQueueListenerManager.Setup(x => x.StartListener());
+            Subject.MessageQueueListenerManager = mockMessageQueueListenerManager.Object;
+
+            Subject.StartMessageQueueListener();
+
+            mockMessageQueueListenerManager.Verify(x => x.StartListener(), Times.Once);
         }
 
         [Test]
@@ -75,6 +92,38 @@ namespace IntegrationEngine.Tests
             Subject.Dispose();
 
             mockWebApiApplication.Verify(x => x.Dispose(), Times.Once);
+        }
+
+        [Test]
+        public void ShouldRegisterIntegrationPoints()
+        {
+            Subject.EngineConfiguration = new EngineConfiguration();
+            Subject.LoadConfiguration();
+
+            Subject.RegisterIntegrationPoints();
+
+            var container = Subject.Container;
+            container.Resolve<IMailConfiguration>("DefaultMail");
+            container.Resolve<IMailConfiguration>("FooMailClient");
+            container.Resolve<IElasticsearchConfiguration>("DefaultElasticsearch");
+            container.Resolve<IRabbitMQConfiguration>("DefaultRabbitMQ");
+            container.Resolve<IJsonServiceConfiguration>("ExampleJsonService");
+        }
+
+        [Test]
+        public void ShouldRegisterIntegrationJobs()
+        {
+            var assembliesWithJobs = new List<Assembly> { typeof(Program).Assembly };
+            Subject.IntegrationJobTypes = Subject.ExtractIntegrationJobTypesFromAssemblies(assembliesWithJobs);
+            Subject.LoadConfiguration();
+            Subject.RegisterIntegrationPoints();
+
+            Subject.RegisterIntegrationJobs();
+
+            var container = Subject.Container;
+            container.Resolve<CarMailMessageJob>();
+            container.Resolve<CarReportJob>();
+            container.Resolve<SampleSqlReportJob>();
         }
     }
 }