+++ IntegrationEngine/Api/WebApiApplication.cs
@@ -1,6 +1,11 @@
 ﻿using IntegrationEngine.Core.Configuration;
 using Microsoft.Owin.Hosting;
+using Microsoft.Practices.Unity;
+using Owin;
 using System;
+using System.Linq;
+using System.Web.Http;
+using System.Web.Http.Cors;
 
 namespace IntegrationEngine.Api
 {
@@ -8,14 +13,30 @@ namespace IntegrationEngine.Api
     {
         public IDisposable webApi { get; set; }
         public WebApiConfiguration WebApiConfiguration { get; set; }
+        public ContainerResolver ContainerResolver { get; set; }
 
         public WebApiApplication()
-        { }
+        {}
+
+        public HttpConfiguration HttpConfigurationFactory()
+        {
+            var  config = new HttpConfiguration();
+            config.DependencyResolver = ContainerResolver;
+            if (WebApiConfiguration != null && WebApiConfiguration.Origins.Any())
+                config.EnableCors(new EnableCorsAttribute(string.Join(",", WebApiConfiguration.Origins), "*", "*"));
+            config.Routes.MapHttpRoute(
+                name: "DefaultApi",
+                routeTemplate: "api/{controller}/{id}",
+                defaults: new { id = RouteParameter.Optional }
+            );
+            config.MapHttpAttributeRoutes();
+            return config;
+        }
 
         public void Start()
         {
             var baseAddress = (new UriBuilder("http", WebApiConfiguration.HostName, WebApiConfiguration.Port)).Uri.AbsoluteUri;
-            webApi = WebApp.Start<WebApiStartup>(baseAddress);
+            webApi = WebApp.Start(baseAddress, x => x.UseWebApi(HttpConfigurationFactory()));
         }
 
         public void Dispose()