+++ IntegrationEngine/Scheduler/SimpleTrigger.cs
@@ -7,7 +7,6 @@ namespace IntegrationEngine.Scheduler
     public class SimpleTrigger : IntegrationEngine.Model.SimpleTrigger, IntegrationEngine.Model.IHasParameters
     {
         [Required]
-        [JobType]
         public override string JobType { get; set; }
         [Range(0, 1)]
         public override int StateId { get; set; }