+++ IntegrationServer/IntegrationPoints/ExampleJsonService.cs
@@ -1,0 +1,13 @@
+﻿using IntegrationEngine.Core.Configuration;
+using IntegrationEngine.Core.ServiceStack;
+
+namespace IntegrationServer.IntegrationPoints
+{
+    public class ExampleJsonService : JsonServiceClientAdapter
+    {
+        public ExampleJsonService(IJsonServiceConfiguration jsonServiceConfiguration)
+        {
+            JsonServiceConfiguration = jsonServiceConfiguration;
+        }
+    }
+}