+++ IntegrationServer/Program.cs
@@ -2,13 +2,14 @@
 using System.IO;
 using System.Reflection;
 using System.ServiceProcess;
+using IntegrationEngine;
 
-namespace IntegrationEngine.ConsoleHost
+namespace IntegrationServer
 {
     public class Program
     {
         public const string ServiceName = "InEngine.NET Server";
-        public static EngineHost EngineHosts { get; set; }
+        public static EngineHost EngineHost { get; set; }
         public static void Main(string[] args)
         {
             var isRunningUnderMono = Type.GetType("Mono.Runtime") != null;
@@ -30,13 +31,13 @@ namespace IntegrationEngine.ConsoleHost
 
         private static void Start(string[] args)
         {
-            EngineHosts = new EngineHost(typeof(Program).Assembly);
-            EngineHosts.Initialize();
+            EngineHost = new EngineHost(typeof(Program).Assembly);
+            EngineHost.Initialize();
         }
 
         private static void Stop()
         {
-            EngineHosts.Dispose();
+            EngineHost.Dispose();
         }
 
         public class Service : ServiceBase