+++ configuration/SharedAssemblyInfo.cs
@@ -2,8 +2,8 @@
 
 // Assembly Info that is shared across the product
 [assembly: AssemblyProduct("InEngine.NET")]
-[assembly: AssemblyVersion("2.0.1.*")]
-[assembly: AssemblyInformationalVersion("2.0.1")]
+[assembly: AssemblyVersion("2.0.2.*")]
+[assembly: AssemblyInformationalVersion("2.0.2")]
 
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Ethan Hann")]